Remove *.gen.ts and make node instances as pure objects
