+++ src/nodes/location.ts
@@ -1,0 +1,23 @@
+export interface Position {
+  offset: number
+  line: number
+  column: number
+}
+
+export interface Location {
+  start: Position
+  end: Position
+}
+
+export const buildEmptryLocation = (): Location => ({
+  start: {
+    offset: NaN,
+    line: NaN,
+    column: NaN,
+  },
+  end: {
+    offset: NaN,
+    line: NaN,
+    column: NaN,
+  },
+})
