+++ src/nodes/program.ts
@@ -1,0 +1,7 @@
+import { BaseNode } from '.'
+import { Statement } from './statement'
+
+export interface Program extends BaseNode {
+  type: 'Program'
+  body: Array<Statement>
+}
